Technologies:  HTML CSS JavaScript Solidity
This project is a decentralized fundraising platform that allows individuals and organizations to raise funds for various causes without the need for intermediaries. The platform is built using HTML, CSS, JavaScript, and Solidity smart contracts. The front-end of the platform enables users to create and manage their fundraising campaigns, view donor information, and monitor the progress of their campaigns.